requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
idna==3.10
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
requests==2.32.4
rsa==4.9.1
six==1.17.0
urllib3==2.4.0
python-dotenv==1.0.1
pandas==2.2.1
google-generativeai==0.4.1
colorama==0.4.6
